+++ changelog.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.13.0] 2023-03-25
 
-- Support quoted string options
+- Support quoted string options, quote chars are ", ' or `
 
 ## [0.12.0] 2022-04-14